created validate for delete of profile

+++ app/Http/Controllers/Member/MemberController.php
@@ -52,8 +52,10 @@ class MemberController extends Controller
 
     public function destroy(Member $member): RedirectResponse
     {
-        $member->user->delete();
-        $member->delete();
+        if (!$member->user->super) {
+            $member->user->delete();
+            $member->delete();
+        }
 
         return to_route('members.index');
     }
